check for existing binary before downloading

+++ src/lib.rs
@@ -40,24 +40,38 @@ pub fn init() -> HugoBuilder {
     let url = format!("https://github.com/gohugoio/hugo/releases/download/v{VERSION}/hugo_extended_{VERSION}_{ARCH}.tar.gz");
     let out_dir = std::env::var("OUT_DIR").unwrap();
     let out_path = Path::new(&out_dir);
-    let result = reqwest::blocking::get(url).unwrap();
-    let bytes = result.bytes().unwrap();
-    let decompressor = GzDecoder::new(&bytes[..]);
-    let mut archive = Archive::new(decompressor);
     let mut binary_name = out_path.join("hugo");
-    for entry in archive.entries().unwrap() {
-        let mut file = entry.unwrap();
-        let file_path = file.path().unwrap();
-        let is_binary = file_path.starts_with("hugo");
-        let target_file_name = out_path.join(&file_path);
-        let mut bytes: Vec<u8> = vec![];
-        _ = file.read_to_end(&mut bytes).unwrap();
-        let mut local_file = File::create(target_file_name.clone()).unwrap();
-        local_file.write_all(&bytes).unwrap();
-        if is_binary {
-            binary_name = out_path.join(target_file_name.clone());
-            #[cfg(not(target_os = "windows"))]
-            fix_permissions(&local_file);
+
+    // check for already downloaded binary
+    let mut binary_exists = false;
+    let result = out_path.read_dir().expect("reading OUT_DIR");
+    for file in result {
+        let entry = file.unwrap();
+        if entry.file_name().to_string_lossy().contains("hugo") {
+            binary_exists = true;
+            binary_name = entry.path();
+        }
+    }
+    if !binary_exists {
+        // download fresh binary
+        let result = reqwest::blocking::get(url).unwrap();
+        let bytes = result.bytes().unwrap();
+        let decompressor = GzDecoder::new(&bytes[..]);
+        let mut archive = Archive::new(decompressor);
+        for entry in archive.entries().unwrap() {
+            let mut file = entry.unwrap();
+            let file_path = file.path().unwrap();
+            let is_binary = file_path.starts_with("hugo");
+            let target_file_name = out_path.join(&file_path);
+            let mut bytes: Vec<u8> = vec![];
+            _ = file.read_to_end(&mut bytes).unwrap();
+            let mut local_file = File::create(target_file_name.clone()).unwrap();
+            local_file.write_all(&bytes).unwrap();
+            if is_binary {
+                binary_name = out_path.join(target_file_name.clone());
+                #[cfg(not(target_os = "windows"))]
+                fix_permissions(&local_file);
+            }
         }
     }
     HugoBuilder {
